Projeto ajustado
Corrigi o projeto no meu computador e agora está rodando, pesquisem em como usar o git no eclipse, para não ficar dando problemas no projeto, que assim não precisa importar pasta, mas sim o projeto do git.

+++ src/pessoas/Pessoa.java
@@ -7,7 +7,7 @@ public class Pessoa {
 
 	private String nomeCompleto;
 	private Calendar dataNascimento;
-	private int cpf;
+	private long cpf;
 	private String endereco;
 	private int telefone;
 	private String tipo;
@@ -107,7 +107,7 @@ public class Pessoa {
 	}
 
 
-	public int getCpf() {
+	public long getCpf() {
 		return this.cpf;
 		
 	}
